refactor: udpate query
