+++ src/store/modules/dashboard/dashboard-data-layout.ts
@@ -140,7 +140,7 @@ const mutations: MutationTree<State> = {
             t: 'Title',
             h: 250,
             version: '1.0',
-            currentService: param.currentService,
+            currentService: '',
             entityType: 'Service',
             independentSelector: true,
             metricType: 'UNKNOWN',
@@ -151,7 +151,7 @@ const mutations: MutationTree<State> = {
             t: 'Title',
             h: 250,
             version: '1.0',
-            currentService: param.currentService,
+            currentService: '',
             entityType: 'Service',
             independentSelector: true,
             metricType: 'UNKNOWN',
