+++ src/store/modules/dashboard/dashboard-data-query.ts
@@ -40,6 +40,7 @@ const actions: ActionTree<State, any> = {
     if (!config.metricName) {
       return;
     }
+    // remove the space at the beginning and end of the string
     const metricNames = config.metricName.split(',').map((item: string) => item.replace(/^\s*|\s*$/g, ''));
     const metricLabels = (config.metricLabels || '').split(',').map((item: string) => item.replace(/^\s*|\s*$/g, ''));
     const currentServiceId = config.independentSelector ? config.currentService : currentService.key;
