+++ src/store/modules/dashboard/dashboard-option.ts
@@ -232,6 +232,19 @@ const actions: ActionTree<State, any> = {
         return res.data.data.getServiceInstances;
       });
   },
+  TYPE_METRICS(context, params: { name: string }) {
+    const metricNames = (params.name || '').split(',').map((item: string) => item.replace(/^\s*|\s*$/g, ''));
+    return Promise.all(
+      metricNames.map((item: string) => {
+        return graph
+          .query('queryTypeOfMetrics')
+          .params({ name: item })
+          .then((res: AxiosResponse) => {
+            return res.data.data;
+          });
+      }),
+    );
+  },
 };
 
 export default {
